requirements.txt
addict==2.4.0
aiofiles==23.2.1
albumentations==1.4.2
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
gitdb==4.0.11
gradio==4.22.0
gradio_client==0.13.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
loralib==0.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmengine==0.10.3
networkx==3.1
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.9.15
packaging==24.0
pandas==2.0.3
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
prettytable==3.10.0
protobuf==4.25.3
psutil==5.9.8
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.28.1
rich==13.7.1
rotary-embedding-torch==0.3.6
rpds-py==0.18.0
ruff==0.3.4
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.36.3
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.9.16
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
typer==0.10.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.13
uvicorn==0.29.0
wandb==0.16.4
wcwidth==0.2.13
websockets==11.0.3
yapf==0.40.2
zipp==3.18.1
